refactor: Change the name from 'config' to 'properties'
